Bootstrap stats client and mapnik fonts in new test client
deletes old test client
